Test CRUD on comments box

+++ parentcodeapp/views.py
@@ -126,16 +126,17 @@ def comment_delete(request, slug, comment_id):
     """
     view to delete comment
     """
-    comment = get_object_or_404(Comment, pk=comment_id)
+    if request.method == 'POST':
+        comment = get_object_or_404(Comment, pk=comment_id)
 
-    if comment.author == request.user:
-        comment.delete()
-        messages.add_message(request, messages.SUCCESS, 'Comment deleted!')
-    else:
-        messages.add_message(
-            request,
-            messages.ERROR,
-            'You can only delete your own comments!'
-        )
-
+        if comment.author == request.user:
+            comment.delete()
+            messages.add_message(request, messages.SUCCESS, 'Comment deleted!')
+        else:
+            messages.add_message(
+                request,
+                messages.ERROR,
+                'You can only delete your own comments!'
+            )
+    
     return redirect('post_detail', slug=slug)
